Fix broken rtimer for native
`rtimer_arch_schedule/1`  would cast clock_second as `unsigned short` instead
of `unsigned int` causing scheduling to set wrong interval.

Example:

```c
rtimer_set(&timer, RTIMER_NOW() + CLOCK_SECOND, 0, callback, null);

/* with the original code the following would be ouputted from `rtimer-arch.c`:
 * > rtimer_arch_schedule time 46487236 46466024 in 46466.24000 seconds
 *
 * With fix the following, corrected version:
 * > rtimer_arch_schedule time 46463604 1000 in 1.0 seconds
 */

+++ arch/cpu/native/rtimer-arch.c
@@ -77,13 +77,13 @@ rtimer_arch_schedule(rtimer_clock_t t)
   struct itimerval val;
   rtimer_clock_t c;
 
-  c = t - (unsigned short)clock_time();
+  c = t - clock_time();
   
   val.it_value.tv_sec = c / CLOCK_SECOND;
   val.it_value.tv_usec = (c % CLOCK_SECOND) * CLOCK_SECOND;
 
-  PRINTF("rtimer_arch_schedule time %u %u in %d.%d seconds\n", t, c, val.it_value.tv_sec,
-      val.it_value.tv_usec);
+  PRINTF("rtimer_arch_schedule time %"PRIu32 " %"PRIu32 " in %ld.%ld seconds\n",
+         t, c, (long)val.it_value.tv_sec, (long)val.it_value.tv_usec);
 
   val.it_interval.tv_sec = val.it_interval.tv_usec = 0;
   setitimer(ITIMER_REAL, &val, NULL);
